# helmsman-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local Kubernetes cluster management API. Connects to whichever cluster your kubeconfig points at.",
        "title": "Helmsman API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/contexts/{ctx}/resources/{resource}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "List resources (server-side table)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context name or _current",
                        "name": "ctx",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource (e.g. pods, deployments.apps)",
                        "name": "resource",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.TablePayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.RowObject": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "handler.TableColumn": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "handler.TablePayload": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TableColumn"
                    }
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TableRow"
                    }
                }
            }
        },
        "handler.TableRow": {
            "type": "object",
            "properties": {
                "cells": {
                    "type": "array",
                    "items": {}
                },
                "object": {
                    "$ref": "#/definitions/handler.RowObject"
                }
            }
        }
    }
}